requirements.txt
langchain==0.1.0
langchain-community==0.0.20
praw==7.7.1
yfinance==0.2.36
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
streamlit==1.30.0
pandas==2.1.4
plotly==5.18.0
textblob==0.17.1
wordcloud==1.9.3
alpha_vantage 